+++ packages/ui-vue/src/components/select-content.vue
@@ -8,7 +8,7 @@ const props = defineProps<{ class?: string }>()
 <template>
   <SelectPortal>
     <SelectContent v-bind="props" :class="cn('relative z-50 max-h-96 min-w-[8rem] overflow-hidden rounded-md border bg-popover text-popover-foreground shadow-md data-[state=open]:animate-in data-[state=closed]:animate-out', props.class)">
-      <SelectViewport class="p-1"><slot /></SelectViewport>
+      <SelectViewport class="p-1 scroll-py-1"><slot /></SelectViewport>
     </SelectContent>
   </SelectPortal>
 </template>